allow use of import_pillows without gevent patch
